requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==5.1.2
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-tinymce==3.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==21.0.0
idna==3.10
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
Pillow==9.5.0
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
PySnooper==1.2.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reindent==3.5.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
